refactor: rename submissionModel --> taskModel
